create packages for better overview, moved contentcard to separate file
