+++ frontend/lib/verification/view/verification_info_text.dart
